pyproject.toml
[project]
name = "prox-automigrate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paramiko>=4.0.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["prox_automigrate"]

[project.scripts]
prox-am = "prox_automigrate.scripts.start:main"
